adjustment of source code organization
